Wrap per-array inputs in a list
Adds a decorator (@list_or_scalar) used to specify which parameters
should be lists of the same length as PVSystem._arrays. The decorator
handles validation by transforming non-list values to singleton lists
then comparing to the length of the _arrays field on the PVSystem
instance passed as the first parameter.

Needs tests with multiple Arrays, but must wait until we have an API for
adding multiple Arrays to a PVSystem.